==> 1_Data_Preprocess.ipynb <==
from pathlib import Path

import pandas as pd

from src import preprocess

%load_ext autoreload
%autoreload 2

DATA_DIR = "./data/"
CLINICAL_NOTES_FILE = DATA_DIR + "ClinNotes.csv"
MEDICAL_CONCEPTS_FILE = DATA_DIR + "MedicalConcepts.csv"

PROCESSDED_DATA_DIR = './processed_data/'
PROCESSED_CLINICAL_NOTES_FILE = PROCESSDED_DATA_DIR + "ClinNotes.csv"

df_clinical = pd.read_csv(CLINICAL_NOTES_FILE)

# Data Understanding

Firstly let's take a look at the csv format.

df_clinical.head()

print('The dataset contains {} records.'.format(len(df_clinical)))

Let's take a peek at one specific clinical note.

df_clinical['notes'][0]

We can observe the data contains more commas than needed. I think this is due to some data normalization steps that are done to collapse properly formatted clinical notes. Let's replace the comma with newline to check out.

preprocess.print_note(df_clinical['notes'][0])

Basically we can see the format looks nicer in this way despite we match all the commas.

For data understanding, we observe that this dataset containes a small collection (around 800) of clinical notes. It also comes with a extra column describing which category the note is coming from. For the format of the clinical note, we need to perform some extra data cleaning steps. And there is also one notable thing is that the notes contain various numbers, which will impose a very big challenge to vectorization because generally speaking numbers are noise in text without any useful semantic meaning. Some sources of the number are dates, time duration, length or width, blood pressure, medicine concentration and etc.

# Data Preprocess

The clinical notes are written in different formats, and it is hard to parse them in a unified way. But most of them follow a standard that one note is segregated into several sections, each section has an all-caps title and a paragraph of content. In this project, I would like to only extract the content of the notes and discard the titles. The main considerations are as follow:
1. The content of each section is quite informative and can be considered as it already covers the semantic meaning of the title.
2. We want the vectorization method to be generalized and not all clinical notes are in the this format. Including the titles may give those word frequency based method some advantages. And extrating the content will make our vertorization study more robust and generalized to all scenarios.

Let's use the regular expression and some pre-defined rules to match the title and try with some examples below.

preprocess.match_titles(df_clinical['notes'][0])

preprocess.match_titles(df_clinical['notes'][3])

preprocess.match_titles(df_clinical['notes'][9])

Then we can implement our content extraction method, the basic logic is to firstly extract the titles of the note, and then we only keep the fraction of the notes that are not the titles. Let's use the same notes as examples below.

preprocess.extract_content(df_clinical['notes'][0])

preprocess.extract_content(df_clinical['notes'][3])

Here I did some extra cleaning to also remove the section content if the section title contains 'TIME' or 'DATE', this is mainly because I want to reduce the numbers in the content, which can be challenging for vectorization.

preprocess.extract_content(df_clinical['notes'][9])

Besides removing the dates and times, I also want to remove the numbered index from the content. We can see the outcome of the function below.

preprocess.remove_number_index(df_clinical['notes'][10])

Some extra data cleaning steps include removing extra commas and removing redundant white spaces. After the data cleaning steps, I found some notes become empty strings. For simplicity, I just fill it with some dummy values here. In a more serious context, we may consider to remove those empty notes. With all the data preprocessing and cleaning functions in place, we can batch process our clinical notes data.

df_clinical_processed = df_clinical.copy()
df_clinical_processed['notes'] = df_clinical_processed['notes'].apply(lambda s: \
    preprocess.fill_empty_note(                                                                  
    preprocess.remove_redundant_whitespace( \
    preprocess.remove_commas( \
    preprocess.remove_number_index( \
    preprocess.extract_content(s) \
)))))

Let's also save them for reusing in later notebooks.

Path(PROCESSDED_DATA_DIR).mkdir(parents=True, exist_ok=True)
df_clinical_processed.to_csv(PROCESSED_CLINICAL_NOTES_FILE, index=False)
