==> 7_Similarity_Ranking_Task.ipynb <==
import numpy as np
import pandas as pd

from src import preprocess, evaluate

%load_ext autoreload
%autoreload 2

DATA_DIR = "./data/"
CLINICAL_NOTES_FILE = DATA_DIR + "ClinNotes.csv"
MEDICAL_CONCEPTS_FILE = DATA_DIR + "MedicalConcepts.csv"

PROCESSDED_DATA_DIR = './processed_data/'
TFIDF_VECTOR_FILE = PROCESSDED_DATA_DIR + 'Tfidf_vector.npy'
TFIDF_EXTENDED_VECTOR_FILE = PROCESSDED_DATA_DIR + 'Tfidf_extended_vector.npy'
BIOWORDVEC_VECTOR_FILE = PROCESSDED_DATA_DIR + 'BioWordVec_vector.npy'
CLINICAL_BERT_VECTOR = PROCESSDED_DATA_DIR + 'Clinical_Bert_vector.npy'

# Use Case: Similarity
Once we have the document vectors, we can measure the similarity between two vectors. This is very helpful in searching and ranking. The use cases can be given a user query or a clinical note, to find the most relevant ones from our database. I use the cosine similarity as the metric to avoid favoring larger norms of the vector. Below I show the use case of given a query clinical note, to find the top-k most relevant ones.

Since we don't have labeled data for the relevancy and I do not have the domain knowledge to give a qualitative analysis, I will skip the evaluation part for this task. I will try to interpret the examples below based on the vectorization techniques.

df_clinical = pd.read_csv(CLINICAL_NOTES_FILE)

vectors_tfidf = np.load(TFIDF_VECTOR_FILE)
vectors_tfidf_extended = np.load(TFIDF_EXTENDED_VECTOR_FILE)
vectors_biowordvec = np.load(BIOWORDVEC_VECTOR_FILE)
vectors_clinical_bert = np.load(CLINICAL_BERT_VECTOR)

## TF-IDF
We can observe the matching basically depends on the overlapping of words. The result of extended version of tf-idf vector shows some minor difference to the plain one.

evaluate.print_topk_most_similar_notes(df_clinical, vectors_tfidf, 0, 3)

evaluate.print_topk_most_similar_notes(df_clinical, vectors_tfidf_extended, 0, 3)

## Word Vector Aggregation
We can observe the result is quite similar to the TF-IDF vectorization. This is expected as the overlapping of words will result in a similar aggregated word vector.

evaluate.print_topk_most_similar_notes(df_clinical, vectors_biowordvec, 0, 3)

## Language Model based Vectorization
We can observe the results are quite different from above two methods and there is likely a favor for long documents.

evaluate.print_topk_most_similar_notes(df_clinical, vectors_clinical_bert, 0, 3)
